test(forms): move ngSubmit to built-in Angular examples #756
