+++ docs/articles/api/ngMocks/output.md
@@ -19,5 +19,5 @@ const outputEmitter = ngMocks.output(debugElement, 'update');
 const outputEmitter = ngMocks.output('app-component', 'update');
 ```
 
-For `ngSubmit`, see [testing Angular forms](/guides/libraries/angular-forms.md)
+For `ngSubmit`, see [testing Angular forms](/guides/ng-submit.md)
 for examples of emitting the output and triggering a native form submission.
