+++ docs/articles/extra/mock-form-controls.md
@@ -30,7 +30,7 @@ it supports both `FormsModule` and `ReactiveFormsModule`:
 
 ## Related tools
 
-- [Testing form submission and `ngSubmit`](/guides/libraries/angular-forms.md)
+- [Testing form submission and `ngSubmit`](/guides/ng-submit.md)
 - [`ngMocks.change()`](/api/ngMocks/change.md)
 - [`ngMocks.touch()`](/api/ngMocks/touch.md)
