+++ examples/TestNgSubmit/reactive.spec.ts
@@ -3,62 +3,71 @@ import {
   Component,
   NgModule,
 } from '@angular/core';
-import { FormsModule, NgForm } from '@angular/forms';
+import {
+  FormControl,
+  FormGroup,
+  FormGroupDirective,
+  ReactiveFormsModule,
+} from '@angular/forms';
+
 import { isMockOf, MockBuilder, MockRender, ngMocks } from 'ng-mocks';
 
 @Component({
-  selector: 'target-ng-submit-template-driven',
-  standalone: false,
+  selector: 'target-ng-submit-reactive',
+  ['standalone' as never /* TODO: remove after upgrade to a14 */]: false,
   template: `
-    <form (ngSubmit)="save(value, $event)">
-      <input
-        name="name"
-        [(ngModel)]="value"
-        [ngModelOptions]="{ updateOn: 'submit' }"
-      />
+    <form
+      [formGroup]="form"
+      (ngSubmit)="save(form.controls.name.value, $event)"
+    >
+      <input formControlName="name" />
       <button type="submit" [disabled]="disabled">Save</button>
     </form>
   `,
 })
 class TargetComponent {
   public disabled = false;
-  public value = 'initial';
+  public readonly form = new FormGroup({
+    name: new FormControl('initial', {
+      updateOn: 'submit',
+    }),
+  });
   public readonly submissions: Array<{
-    value: string;
+    value: string | null;
     event: Event;
   }> = [];
 
-  public save(value: string, event: Event): void {
+  public save(value: string | null, event: Event): void {
     this.submissions.push({ value, event });
   }
 }
 
 @NgModule({
   declarations: [TargetComponent],
-  imports: [FormsModule],
+  imports: [ReactiveFormsModule],
 })
 class TargetModule {}
 
 // @see https://github.com/help-me-mom/ng-mocks/issues/756
-describe('ng-submit:template-driven', () => {
+describe('TestNgSubmit:reactive', () => {
   describe('real', () => {
     beforeEach(() =>
-      MockBuilder(TargetComponent, TargetModule).keep(FormsModule),
+      MockBuilder(TargetComponent, TargetModule).keep(
+        ReactiveFormsModule,
+      ),
     );
 
-    it('submits pending values and forwards the submit event', async () => {
-      const fixture = MockRender(TargetComponent);
-      await fixture.whenStable();
-      const component = fixture.point.componentInstance;
-      const form = ngMocks.findInstance(NgForm);
+    it('submits pending values and forwards the submit event', () => {
+      const component =
+        MockRender(TargetComponent).point.componentInstance;
+      const form = ngMocks.findInstance(FormGroupDirective);
 
       expect(form.submitted).toBe(false);
-      expect(form.value).toEqual({ name: 'initial' });
+      expect(component.form.value).toEqual({ name: 'initial' });
       expect(component.submissions).toEqual([]);
 
       ngMocks.change('input', 'updated');
-      expect(component.value).toBe('initial');
-      expect(form.value).toEqual({ name: 'initial' });
+      expect(component.form.value).toEqual({ name: 'initial' });
       expect(component.submissions).toEqual([]);
 
       // A native submit synchronizes updateOn: 'submit' controls
@@ -67,19 +76,17 @@ describe('ng-submit:template-driven', () => {
       ngMocks.trigger('form', event);
 
       expect(form.submitted).toBe(true);
-      expect(form.value).toEqual({ name: 'updated' });
-      expect(component.value).toBe('updated');
+      expect(component.form.value).toEqual({ name: 'updated' });
       expect(component.submissions).toEqual([
         { value: 'updated', event },
       ]);
       expect(event.defaultPrevented).toBe(true);
     });
 
-    it('submits through the native submit button', async () => {
-      const fixture = MockRender(TargetComponent);
-      await fixture.whenStable();
-      const component = fixture.point.componentInstance;
-      const form = ngMocks.findInstance(NgForm);
+    it('submits through the native submit button', () => {
+      const component =
+        MockRender(TargetComponent).point.componentInstance;
+      const form = ngMocks.findInstance(FormGroupDirective);
 
       ngMocks.change('input', 'updated');
       // eslint-disable-next-line es-x/no-array-prototype-find -- ngMocks.find is not Array.find.
@@ -88,7 +95,7 @@ describe('ng-submit:template-driven', () => {
       button.click();
 
       expect(form.submitted).toBe(true);
-      expect(form.value).toEqual({ name: 'updated' });
+      expect(component.form.value).toEqual({ name: 'updated' });
       expect(component.submissions.length).toBe(1);
       expect(component.submissions[0].value).toBe('updated');
       expect(component.submissions[0].event.type).toBe('submit');
@@ -97,11 +104,10 @@ describe('ng-submit:template-driven', () => {
       );
     });
 
-    it('does not submit through a disabled button', async () => {
+    it('does not submit through a disabled button', () => {
       const fixture = MockRender(TargetComponent);
-      await fixture.whenStable();
       const component = fixture.point.componentInstance;
-      const form = ngMocks.findInstance(NgForm);
+      const form = ngMocks.findInstance(FormGroupDirective);
       component.disabled = true;
       fixture.point.injector.get(ChangeDetectorRef).markForCheck();
       fixture.detectChanges();
@@ -114,16 +120,14 @@ describe('ng-submit:template-driven', () => {
       button.click();
 
       expect(component.submissions).toEqual([]);
-      expect(component.value).toBe('initial');
-      expect(form.value).toEqual({ name: 'initial' });
+      expect(component.form.value).toEqual({ name: 'initial' });
       expect(form.submitted).toBe(false);
     });
 
-    it('emits the output without submitting pending values', async () => {
-      const fixture = MockRender(TargetComponent);
-      await fixture.whenStable();
-      const component = fixture.point.componentInstance;
-      const form = ngMocks.findInstance(NgForm);
+    it('emits the output without submitting pending values', () => {
+      const component =
+        MockRender(TargetComponent).point.componentInstance;
+      const form = ngMocks.findInstance(FormGroupDirective);
       ngMocks.change('input', 'updated');
 
       const event = ngMocks.event('submit');
@@ -132,8 +136,7 @@ describe('ng-submit:template-driven', () => {
       expect(component.submissions).toEqual([
         { value: 'initial', event },
       ]);
-      expect(component.value).toBe('initial');
-      expect(form.value).toEqual({ name: 'initial' });
+      expect(component.form.value).toEqual({ name: 'initial' });
       expect(form.submitted).toBe(false);
       expect(event.defaultPrevented).toBe(false);
     });
@@ -142,20 +145,24 @@ describe('ng-submit:template-driven', () => {
   describe('mock', () => {
     beforeEach(() => MockBuilder(TargetComponent, TargetModule));
 
-    it('binds ngModel and ngSubmit', () => {
+    it('binds formGroup and ngSubmit', () => {
       const component =
         MockRender(TargetComponent).point.componentInstance;
       expect(
-        isMockOf(ngMocks.findInstance(NgForm), NgForm, 'd'),
+        isMockOf(
+          ngMocks.findInstance(FormGroupDirective),
+          FormGroupDirective,
+          'd',
+        ),
       ).toBe(true);
-      expect(ngMocks.input('input', 'ngModel')).toBe('initial');
+      expect(ngMocks.input('form', 'formGroup')).toBe(component.form);
+      expect(ngMocks.input('input', 'formControlName')).toBe('name');
       expect(component.submissions).toEqual([]);
 
-      ngMocks.output('input', 'ngModelChange').emit('updated');
+      component.form.setValue({ name: 'updated' });
       const event = ngMocks.event('submit');
       ngMocks.output('form', 'ngSubmit').emit(event);
 
-      expect(component.value).toBe('updated');
       expect(component.submissions).toEqual([
         { value: 'updated', event },
       ]);
